+++ R/epmc_details.r
@@ -61,7 +61,7 @@ epmc_details <- function(ext_id = NULL, data_src = "med") {
     q <- list(query = paste0("PMCID:", ext_id), format = "json",
             resulttype = "core")
   } else {
-    q <- list(query = paste0("ext_id:", ext_id, " src:", data_src),
+    q <- list(query = paste0("ext_id:", ext_id, "%20src:", data_src),
               format = "json", resulttype = "core")
   }
   doc <- rebi_GET(path = path, query = q)