+++ R/epmc_hits.r
@@ -15,8 +15,7 @@
 
 epmc_hits <- function(query = NULL, ...) {
   # check
-  if (is.null(query))
-    stop("No query provided")
+  query <- transform_query(query)
   path = paste0(rest_path(), "/search")
   doc <- rebi_GET(path, query = list(query = query, format = "json", ...))
   hit_count <- doc$hitCount