+++ R/epmc_search.r
@@ -10,12 +10,16 @@
 #' @param id_list logical, should only IDs (e.g. PMID) and sources be retrieved
 #'   for the given search terms?
 #' @param limit integer, limit the number of records you wish to retrieve.
-#'   By default, 25 are returned.
+#'   By default, 100 are returned.
 #' @param synonym logical, synonym search. If TRUE, synonym terms from MeSH
 #'  terminology and the UniProt synonym list are queried, too. Disabled by
 #'  default.
+#' @param sort character, sort results by order (\code{asc}, \code{desc}) and
+#'  sort field (e.g. \code{CITED}, \code{P_PDATE}), seperated with a blank.
+#'  For example, sort results  by times cited in descending order:
+#'  \code{sort = "CITED desc"}.
 #' @param verbose	logical, print some information on what is going on.
-#' @return data.frame
+#' @return tibble
 #' @examples \dontrun{
 #' #Search articles for 'Gabi-Kat'
 #' my.data <- epmc_search(query='Gabi-Kat')
@@ -34,55 +38,135 @@
 #' # include mesh and uniprot synonyms in search
 #' my.data <- epmc_search(query = 'aspirin', synonym = TRUE)
 #'
+#' # get 100 most cited atricles from PLOS ONE
+#' epmc_search(query = 'ISSN:	1932-6203', sort = 'CITED desc')
+#'
 #' # print number of records found
 #' attr(my.data, "hit_count")
 #'
 #' }
 #' @export
-
-epmc_search <- function(query = NULL, limit = 25, id_list = FALSE,
-                        synonym = FALSE, verbose = TRUE){
-  # check
-  if (is.null(query))
-    stop("No query provided")
+epmc_search <- function(query = NULL,
+                            id_list = FALSE,
+                            synonym = FALSE,
+                            verbose = TRUE,
+                            limit = 100,
+                            sort = NULL) {
+  query <- transform_query(query)
+  stopifnot(is.logical(c(verbose, id_list, synonym)))
   stopifnot(is.numeric(limit))
-  stopifnot(is.logical(verbose))
-  stopifnot(is.logical(id_list))
-  stopifnot(is.logical(synonym))
-  #  get the correct hit count when mesh and uniprot synonyms are also searched
-  synonym = ifelse(synonym == FALSE, "false", "true")
-  # get results found
-  hit_count <- epmc_hits(query = query, synonym = synonym)
-  #prepare queries
-  queries <- make_queries(hit_count = hit_count, limit = limit, query = query)
-  # if only ids are requested
-  if (id_list == TRUE)
-    queries <-
-    lapply(queries, function(x)
-      c(x, resulttype = "idlist"))
-  # include synonyms from mesh and uniprot
-  if (synonym == TRUE)
-    queries <-
-    lapply(queries, function(x)
-      c(x, synonym = "true"))
-  # get and parse the json from the queries
-  out <- lapply(queries, function(x) {
+  # get the correct hit count when mesh and uniprot synonyms are also searched
+  synonym <- ifelse(synonym == FALSE, "false", "true")
+  page_token <- "*"
+  results <- dplyr::data_frame()
+  out <-
+    epmc_search_(
+      query = query,
+      limit = limit,
+      id_list = id_list,
+      synonym = synonym,
+      verbose = verbose,
+      page_token = page_token,
+      sort = sort
+    )
+  res_chunks <- chunks(limit = limit)
+  # super hacky to control limit, better approach using pageSize param needed
+  hits <- epmc_hits(query, synonym = synonym)
+  limit <- as.integer(limit)
+  limit <- ifelse(hits <= limit, hits, limit)
+  # let's loop over until page max is reached, or until cursor marks are identical
+  i <- 0
+  while (i < res_chunks$page_max) {
+    out <-
+      epmc_search_(
+        query = query,
+        limit = limit,
+        id_list = id_list,
+        synonym = synonym,
+        verbose = verbose,
+        page_token = page_token,
+        sort = sort
+      )
+    if (page_token == out$next_cursor)
+      break
+    i <- i + 1
     if(verbose == TRUE)
-      message(paste0(hit_count, " records found. Retrieving batch ",
-                     x[["page"]]))
-    out <- rebi_GET(path = paste0(rest_path(), "/search"), query = x)
-    plyr::ldply(out$resultList,
-                data.frame,
-                stringsAsFactors = FALSE,
-                .id = NULL)
-  })
-  #combine all into one
-  result <- jsonlite::rbind.pages(out)
-  # remove nested lists from data.frame, get these infos with epmc_details
-  md <- result[, !(names(result) %in% fix_list(result))]
-  # return (thanks to @cstubben)
-  attr(md, "hit_count") <- hit_count
-  md
+      message(paste("Retrieving result page", i))
+    page_token <- out$next_cursor
+    results <- dplyr::bind_rows(results, out$results)
+  }
+  # again, approach needed to use param pageSize instead
+  md <- results[1:limit, ]
+  # return hit counts(thanks to @cstubben)
+  attr(md, "hit_count") <- hits
+  return(md)
 }
 
+#' Get one page of results when searching Europe PubMed Central
+#'
+#' In general, use \code{\link{epmc_search}} instead. It calls this function, calling all
+#' pages within the defined limit.
+#'
+#' @param query character, search query. For more information on how to
+#'   build a search query, see \url{http://europepmc.org/Help}
+#' @param id_list logical, should only IDs (e.g. PMID) and sources be retrieved
+#'   for the given search terms?
+#' @param limit integer, limit the number of records you wish to retrieve.
+#'   By default, 25 are returned.
+#' @param synonym logical, synonym search. If TRUE, synonym terms from MeSH
+#'  terminology and the UniProt synonym list are queried, too. Disabled by
+#'  default.
+#' @param sort character, sort results by order (\code{asc}, \code{desc}) and
+#'  sort field (e.g. \code{CITED}, \code{P_PDATE}), seperated with a blank.
+#'  For example, sort results  by times cited in descending order:
+#'  \code{sort = "CITED desc"}.
+#' @param page_token cursor marking the page
+#'
+#' @param ... further params from \code{\link{epmc_search}}
+#'
+#' @export
+#'
+#' @seealso \link{epmc_search}
+epmc_search_ <-
+  function(query = NULL,
+           limit = 100,
+           id_list = FALSE,
+           synonym = FALSE,
+           page_token = NULL,
+           sort = NULL, ...) {
+    resulttype <- ifelse(id_list == FALSE, "lite", "idlist")
+    # control limit
+    limit <- as.integer(limit)
+    page_size <- ifelse(batch_size() <= limit, batch_size(), limit)
+    #build query
+    args <-
+      list(
+        query = query,
+        format = "json",
+        synonym = synonym,
+        resulttype = resulttype,
+        pageSize = page_size,
+        cursorMark = page_token,
+        sort = val_args(sort)
+      )
+    # call API
+    out <-
+      rebi_GET(path = paste0(rest_path(), "/search"), query = args)
+    # remove nested lists from resulting data.frame, get these infos with epmc_details
+    md <- out$resultList$result
+    if (length(md) == 0) {
+      md <- dplyr::data_frame()
+    } else {
+      md <- md %>%
+        dplyr::select_if(Negate(is.list)) %>%
+        as_data_frame()
+    }
+    list(next_cursor = out$nextCursorMark, results = md)
+  }
+
+val_args <- function(x) {
+  if(!is.null(x))
+    utils::URLencode(x)
+  NULL
+}